Adds functions to PeerManager that enable devices to decide which of two devices will initiate an exchange, and use that function to decide whether to initiate or wait for exchanges from peers.
Use MAC ordering to decide whether to initiate exchange with peer.

Fix errors introduced in rebase.

Added BTSpeakerTest and MacOrderingTests

Added MAC ordering methods to PeerManager and tests for them.

The tests are in their own file because in this branch PeerManagerTest
is commented out because it's broken. They'll be merged into the main
PeerManagerTest.java file later on when the patch that unbreaks PM's
tests hits.

R=jesus@denovogroup.org
BUG=

Review URL: https://de-novo-rietveld.appspot.com/2220001