Adds Wire, a protobuf library, and some .protos for our message formats.
Also adds functionality for Exchange to serialize protobuf objects to
streams and to read those objects off of streams.

To add new .proto files, simply add them to //proto-repo and they'll be
automatically compiled and depended on.

----

Adjust CleartextMessages to hold a list of messages instead of strings.

Tests that show that CleartextMessage and CleartextFriends are
functional.

.proto files in //proto-repo are automatically compiled.

Rules can depend on //proto-repo:protobufs to access the protobufs that
are automatically compiled.

Implemented len/val Write and Read for Messages in Exchange.

Exchange implements lengthValueWrite and lengthValueRead, serialize
Messages to and from input/output streams in a length-delimited fashion.

Added methods to lengthValue encode Messages in Exchange.

Decoding is not yet implemented, although a test exists that manually
decodes.

Added Wire and pre-requistites, and a few tests that demo it.

R=barath@denovogroup.org
BUG=

Review URL: https://de-novo-rietveld.appspot.com/1200001