Fixed the background and the message being added to the message store.
Sends the broadcast and receives it, but does not fully fix the issue. Does not necessarily grab the feed if it has already been generated but the user is on a new fragment.

Have updated the feed so that it has a more spaced out look, with an image in the background and icons that are not yet buttons for saving and resharing.

R=lerner@denovogroup.org
BUG=

Review URL: https://de-novo-rietveld.appspot.com/8180001