NotifyDataSetChanged onResume of Opener. Only register in onResume.
Previously we had the following problem: user goes to "New Post" screen,
the broadcast receiver is unregistered in Opener's onPause. Then, the
message is added to the message store, the broadcast goes out, is missed
by the broadcast receiver, then we return to Opener, where we crash due
to having missed the broadcast (and thus not called
notfiyDataSetChanged()). This fixes it by calling notifyDataSetChanged
on the adapter every time we resume. Thus the broadcast listener only
has to deal with changes that happen while the Opener is foregrounded.

R=jesus@denovogroup.org
BUG=

Review URL: https://de-novo-rietveld.appspot.com/230001