Have rebased the ZXING integration to master
This patch sees the integration of the zxing code, it does not have the button to press to generate a QR code of its own yet and it does not have Rangzen specific URI detection. Some other minor bugs should be fixed, such as the way in which it adds a friend (by activity on result or just activity).

Have class not found error while trying to import ZXING.

R=lerner@denovogroup.org
BUG=

Review URL: https://de-novo-rietveld.appspot.com/7260001