Fixed mipmap/BUCK error. Added feed buttons, no functionality for discard or retweet yet.
Fixed the retweet button, now next to the garbage button in the Favorite tab.

Fixed the problem with the overlaying feeds, it was because of TabHost and how it overlays fragments and not replaces them, so I put a white background up to hide them. Also put a divider for the last element in the list so it can be differentiated from the white space.

Removed Image behind feed, issue with favorite fragment not loading completely? It seems like the messages load in favorites but with the Trust format and messages underneath it. Removed some unused calendar and alarms things that were in Rangzen for some reason. Removed the upvoting and downvoting code.

Added retweet, favorite, and trash icons. No functionality for trash or retweet yet. Changed format of the feed, possible bug is present that doesn't show other feed elements after a save and has a floating constant element at the top in Favorites but I think it is fixed in another patch.

Reformatted feed items with discard, favorites buttons as compound drawables.

Fixed the comments for classes and methods and the constants naming.

Fixed comments, method placements and added constants.

Have added saving of messages, no comments for methods or classes yet. Added new feed adapter class

Merging to master

R=lerner@denovogroup.org
BUG=

Review URL: https://de-novo-rietveld.appspot.com/2270001